moved card cost methods to new util class
